feat(feat-3579): relocate StockSeverity/StockStatusFilter/StockAnalysisSortBy enums to Contracts/
