+++ backend/src/Anela.Heblo.Application/Features/Purchase/Contracts/StockAnalysisSortBy.cs
@@ -1,0 +1,11 @@
+namespace Anela.Heblo.Application.Features.Purchase.Contracts;
+
+public enum StockAnalysisSortBy
+{
+    ProductCode,
+    ProductName,
+    AvailableStock,
+    Consumption,
+    StockEfficiency,
+    LastPurchaseDate
+}
