+++ backend/src/Anela.Heblo.Application/Features/Purchase/Contracts/StockSeverity.cs
@@ -1,0 +1,10 @@
+namespace Anela.Heblo.Application.Features.Purchase.Contracts;
+
+public enum StockSeverity
+{
+    Critical,
+    Low,
+    Optimal,
+    Overstocked,
+    NotConfigured
+}
