+++ backend/src/Anela.Heblo.Application/Features/Purchase/Contracts/StockStatusFilter.cs
@@ -1,0 +1,11 @@
+namespace Anela.Heblo.Application.Features.Purchase.Contracts;
+
+public enum StockStatusFilter
+{
+    All,
+    Critical,
+    Low,
+    Optimal,
+    Overstocked,
+    NotConfigured
+}
